% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CFun2.R
\name{CFun2}
\alias{CFun2}
\title{Title}
\usage{
CFun2(
  drug,
  age,
  gender,
  weight,
  agent,
  country,
  drugEffect,
  year,
  illnesses,
  ptname,
  path
)
}
\arguments{
\item{path}{}
}
\description{
Title
}
