% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFun1.R
\name{FFun1}
\alias{FFun1}
\title{Title}
\usage{
FFun1(
  drug_PS,
  drug_concomitant,
  age,
  gender,
  weight,
  agent,
  country,
  drugEffect,
  year,
  illnesses,
  path
)
}
\arguments{
\item{path}{}
}
\description{
Title
}
