% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FFun2.R
\name{FFun2}
\alias{FFun2}
\title{Title}
\usage{
FFun2(
  drug_PS,
  drug_concomitant,
  age,
  gender,
  weight,
  agent,
  country,
  drugEffect,
  year,
  illnesses,
  path
)
}
\arguments{
\item{path}{}
}
\description{
Title
}
